removed cram conversion, since it is not working on zinc-finger

+++ protocols/CopyToResultsDir.sh
@@ -79,7 +79,10 @@ do
 	rsync -a ${intermediateDir}/${sample}.merged.dedup.bam ${projectResultsDir}/alignment/
 	rsync -a ${intermediateDir}/${sample}.merged.dedup.bam.bai ${projectResultsDir}/alignment/
 	rsync -a ${intermediateDir}/${sample}.merged.dedup.bam.md5 ${projectResultsDir}/alignment/
-	rsync -a ${intermediateDir}/${sample}.merged.dedup.bam.cram ${projectResultsDir}/alignment/
+	if [ -f ${intermediateDir}/${sample}.merged.dedup.bam.cram ] 
+	then
+		rsync -a ${intermediateDir}/${sample}.merged.dedup.bam.cram ${projectResultsDir}/alignment/
+	fi
 
 	printf "."
 done
